# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SonificationParams {
  /** Base frequency in Hz for audio synthesis */
  baseFrequency?: number;
  /** Instrument timbre (sine, sawtooth, square, triangle, fm) */
  timbre?: string;
  /**
     * Pulse rate in BPM for variable stars
     * @nullable
     */
  pulseRate?: number | null;
  harmonics?: number[];
}

export interface Star {
  id: number;
  name: string;
  /** @nullable */
  commonName?: string | null;
  /** O, B, A, F, G, K, or M */
  spectralClass: string;
  /**
     * Stellar subtype (e.g. supergiant, dwarf)
     * @nullable
     */
  subtype?: string | null;
  /**
     * Surface temperature in Kelvin
     * @nullable
     */
  temperature?: number | null;
  /**
     * Mass in solar masses
     * @nullable
     */
  massSolar?: number | null;
  /**
     * Radius in solar radii
     * @nullable
     */
  radiusSolar?: number | null;
  /**
     * Luminosity relative to Sun
     * @nullable
     */
  luminositySolar?: number | null;
  /** Distance from Earth in light years */
  distanceLightYears: number;
  /** Whether the star still exists */
  isAlive: boolean;
  /** How many years ago the light we see left the star */
  transmissionAgeYears: number;
  /**
     * Approximate year of death (negative = BCE)
     * @nullable
     */
  deathYear?: number | null;
  /** @nullable */
  constellation?: string | null;
  /** @nullable */
  nasaId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  generatedImageUrl?: string | null;
  sonificationParams?: SonificationParams;
  description: string;
}

export interface FluxPoint {
  /** Time (BJD or relative days) */
  time: number;
  /** Normalized flux value */
  flux: number;
}

export interface LightCurveData {
  starId: number;
  points: FluxPoint[];
  /** curated or nasa_api */
  source: string;
  /**
     * Pulsation period in days
     * @nullable
     */
  period?: number | null;
}

export interface StarImagery {
  starId: number;
  /**
     * URL of visible light image
     * @nullable
     */
  visibleLight?: string | null;
  /**
     * URL of infrared image
     * @nullable
     */
  infrared?: string | null;
  /**
     * URL of X-ray image
     * @nullable
     */
  xray?: string | null;
  /**
     * URL of AI-generated image
     * @nullable
     */
  generated?: string | null;
}

export interface NarrationRequest {
  /** Additional context about what the user is hearing */
  context?: string;
}

export type CatalogStatsSpectralBreakdown = {[key: string]: number};

export interface CatalogStats {
  totalStars: number;
  deadStars: number;
  aliveStars: number;
  spectralBreakdown: CatalogStatsSpectralBreakdown;
  avgTransmissionAge: number;
  farthestStar: Star;
}

export type ListStarsParams = {
/**
 * Filter by spectral class (O, B, A, F, G, K, M)
 */
spectralClass?: string;
/**
 * Filter by alive/dead status
 */
isAlive?: boolean;
};

export type SearchStarsParams = {
/**
 * Star name query
 */
q: string;
};