# lib/api-zod/src/generated/types/catalogStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogStatsSpectralBreakdown } from './catalogStatsSpectralBreakdown';
import type { Star } from './star';

export interface CatalogStats {
  totalStars: number;
  deadStars: number;
  aliveStars: number;
  spectralBreakdown: CatalogStatsSpectralBreakdown;
  avgTransmissionAge: number;
  farthestStar: Star;
}