# lib/api-zod/src/generated/types/fluxPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */

export interface FluxPoint {
  /** Time (BJD or relative days) */
  time: number;
  /** Normalized flux value */
  flux: number;
}