# lib/api-zod/src/generated/types/lightCurveData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */
import type { FluxPoint } from './fluxPoint';

export interface LightCurveData {
  starId: number;
  points: FluxPoint[];
  /** curated or nasa_api */
  source: string;
  /**
     * Pulsation period in days
     * @nullable
     */
  period?: number | null;
}