# lib/api-zod/src/generated/types/narrationRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */

export interface NarrationRequest {
  /** Additional context about what the user is hearing */
  context?: string;
}