# lib/api-zod/src/generated/types/sonificationParams.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */

export interface SonificationParams {
  /** Base frequency in Hz for audio synthesis */
  baseFrequency?: number;
  /** Instrument timbre (sine, sawtooth, square, triangle, fm) */
  timbre?: string;
  /**
     * Pulse rate in BPM for variable stars
     * @nullable
     */
  pulseRate?: number | null;
  harmonics?: number[];
}