# lib/api-zod/src/generated/types/star.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */
import type { SonificationParams } from './sonificationParams';

export interface Star {
  id: number;
  name: string;
  /** @nullable */
  commonName?: string | null;
  /** O, B, A, F, G, K, or M */
  spectralClass: string;
  /**
     * Stellar subtype (e.g. supergiant, dwarf)
     * @nullable
     */
  subtype?: string | null;
  /**
     * Surface temperature in Kelvin
     * @nullable
     */
  temperature?: number | null;
  /**
     * Mass in solar masses
     * @nullable
     */
  massSolar?: number | null;
  /**
     * Radius in solar radii
     * @nullable
     */
  radiusSolar?: number | null;
  /**
     * Luminosity relative to Sun
     * @nullable
     */
  luminositySolar?: number | null;
  /** Distance from Earth in light years */
  distanceLightYears: number;
  /** Whether the star still exists */
  isAlive: boolean;
  /** How many years ago the light we see left the star */
  transmissionAgeYears: number;
  /**
     * Approximate year of death (negative = BCE)
     * @nullable
     */
  deathYear?: number | null;
  /** @nullable */
  constellation?: string | null;
  /** @nullable */
  nasaId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  generatedImageUrl?: string | null;
  sonificationParams?: SonificationParams;
  description: string;
}