# lib/api-zod/src/generated/types/starImagery.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dead Stars Radio API
 * OpenAPI spec version: 0.1.0
 */

export interface StarImagery {
  starId: number;
  /**
     * URL of visible light image
     * @nullable
     */
  visibleLight?: string | null;
  /**
     * URL of infrared image
     * @nullable
     */
  infrared?: string | null;
  /**
     * URL of X-ray image
     * @nullable
     */
  xray?: string | null;
  /**
     * URL of AI-generated image
     * @nullable
     */
  generated?: string | null;
}